Add lookup by platform & ui
