+++ game_frontend/src/styles/_variables.scss
@@ -1,5 +1,5 @@
 $color-background: white;
 
 
-$main-content-max-width: 900px;
+$main-content-max-width: 500px;
 $main-content-min-width: 350px;
