+++ game_frontend/src/styles/components/games.module.scss
@@ -1,3 +1,14 @@
 .game {
     box-shadow: rgba(60, 64, 67, 0.3) 0px 1px 2px 0px, rgba(60, 64, 67, 0.15) 0px 2px 6px 2px;    padding: 8px;
 }
+
+.container {
+    display: flex;
+    flex-direction: column;
+    margin-left: 20px;
+    padding: 8px 0;
+
+    & a:not(:first-child) {
+        padding-top: 8px;
+    }
+}
