--- game_frontend/src/styles/pages/home.module.scss
@@ -1,10 +1,0 @@
-.container {
-    display: flex;
-    flex-direction: column;
-    margin-left: 20px;
-    padding: 8px 0;
-
-    & a:not(:first-child) {
-        padding-top: 8px;
-    }
-}
