+++ game_manager/backend/models.py
@@ -17,6 +17,7 @@ class Mode(models.Model):
 
 class Platform(models.Model):
     name = models.CharField(max_length=200)
+    url_slug = models.CharField(max_length=200, unique=True)
 
     def __str__(self):
         return f"{self.name}"
