+++ game_manager/scripts_data/test_data.json
@@ -80,7 +80,7 @@
                     "release_date": "2023-05-30",
                     "region": "NA",
                     "publishers": ["Prime Matter"],
-                    "platforms": ["MS-DOS"]
+                    "platforms": ["PC"]
                 }
             ]
         },
